Improved error message for redundant option.
If your function takes a (non-multi-) option, and
the user specifies that option twice, Appeal now
gives you a much better error message.

This is a big improvement, but it's still not perfect.
The error isn't literally about specifying the option
twice, it's attempting to write a kwarg when someone
has already written to it.  So if you run the

	go [--north|--south|--east|--west]

example, and you specify

	% go --north --south

it'll chide you for specifying "--south" twice.