Added Appeal.app_class(), hip new app class tech.
Appeal.app_class() returns a tuple: a decorator for your class,
and a decorator for methods in that class.  The class's __init__ is
used as the global command, and methods decorated with the second
decorator are commands.  Handles instancing the class and binding
the instance for you!  And boy did that implementation get complex.

Here's an example using app_class:

    app = appeal.Appeal()
    app_class, command_method = app.app_class()

    @app_class()
    class MyApp:
        def __init__(self, *, verbose=False):
            print(f"MyApp init {verbose=}")
            self.verbose = verbose

        @command_method()
        def add(self, a, b, c):
            print(f"MyApp add {a=} {b=} {c=} {self.verbose=}")

    app.main()