Fix #5: handle required group missing arguments.
This fixes the issue reported in #5 by Ben Nuttall.
Thanks for the report, Ben!

The problem: if you didn't supply sufficient required
arguments to a command, you got a traceback with a
TypeError.  Ben suggested this is something Appeal
should handle with a better error message.

The specific scenario: when running the simple
"Hello, {name}!" example, which takes exactly one
(required) string argument, if you omit the argument,
you get a TypeError instead of a usage error.

Unsurprisingly in retrospect, this *is* something
Appeal was designed to handle.  It was just a bug.

The bug: in the Charm program generated for the
command, it created an ArgumentGroup representing
the initial, required arguments.  When running
the program, the Charm "interpreter" makes a copy
of this ArgumentGroup object so it can modify it,
tracking the running number of arguments added to
that group.  The bug was, the ArgumentGroup.copy()
method didn't propogate the value of the "optional"
flag.  And since "optional" defaults to True, the
initial ArgumentGroup--which should be required--
was instead marked as optional.  Since it was an
optional group, it was okay for it to not get any
arguments, so Appeal proceeded as if everything
was fine.

The one-line fix, to propogate the "optional" state
to the ArgumentGroup, means that Appeal now generates
a usage error, prints usage to the user, etc.