0.5.2: Fix compatibility with Python 3.11.
As of Python 3.11, inspect.Parameter won't allow a name
that happens to be a keyword.  We don't actually have
functions with parameters that are keywords, but, the
way the Charm compiler is written, it uses an
inspect.Parameter to store the name of "the place where
this converter came from", and the base case for the
recursion is the function itself, and if you use a
lambda for a converter it name is "<lambda>".  Appeal
was stripping the angle brackets, but was still using
"lambda", and inspect.Parameter doesn't like that.

We work around it by attempting to fool with internal
implementation details of inspect.Parameter, but even
if that doesn't work Appeal should still work fine.