New-style option mapping is done!  Progress on #3.
Not done yet: usage is wrong, and a regression test is failing,
and I think it's wrong in some details.  But--

Rewrote when options get mapped in compiled code.  Now they're
provisionally mapped just *before* the first positional argument
in that group.

While I was in there, I cleaned up the compiler code some.
Now it's a class instead of a function.
It's still a bit messy overall... but it's improved.