Undoable converters problem seems totally licked!
Now, the compiler decides whether or not a converter is undoable.
Undoable converters:

    * are class Converter (or a subclass of Converter)
    * are optional arguments, either
         * are VAR_POSITIONAL (*args) or VAR_KEYWORD (**kwargs), or
         * have a default value

The compiler sets an "undoable" attribute of append_to_args opcodes.
The interpreter obeys it.  If the converter is undoable, it appends
it to argument_group.parent_converters.

Then, when the program finishes, it iterates backwards over all
the argument groups.  If the argument group is satisfied and
not laden, it undoes all the undoable converters in it.  Otherwise
it stops processing.

This seems to work perfectly for all my funny boundary-condition
test cases!