Removed "options queue" in favor of delaying.
I used to queue up options in compile_parameter into a list, then
flush the list when it was time to map them.  But it's cleaner
to just shove all the option-mapping code into its own function
and call it at the judicious time.  We only need to map functions
once.