Fix mapping compiler degenerate converter trees.
When compiling a function with Appeal for use with
the mapping compiler, if you have a degenerate tree
like this:

def myfloat(x: float):
    ...
def my_top_fn(a: myfloat, b):
    ...

you want the name you look up in the dictionary to
be "a", not "x", and not the name of the parameter
to "float".  How do we handle that?

Appeal uses the name of the highest parameter in that
subtree that's either the root, or the lowest annotation
in the tree that has more than one parameter.  Functions
that take one value and emit one value "feel like"
inline conversions but maybe don't have interesting
parameter names.